Moved 'figures.backfill' to 'figures.pipeline.backfill'
Moving this because it is a pipeline module. Should have beeing in
'figures.pipeline' when first created
